requirements.txt
streamlit
google-generativeai
chromadb
python-dotenv
pypdf
sentence-transformers
langchain
langchain-google-genai